# Experimental/sonic and buffered write