--- a/go_con/go.mod
+++ b/go_con/go.mod
@@ -1,5 +1,16 @@
 module g.io/related_concurrent
 
-go 1.21.1
+go 1.20
 
-require github.com/goccy/go-json v0.10.2
+require (
+	github.com/bytedance/sonic v1.10.1
+	github.com/goccy/go-json v0.10.2
+)
+
+require (
+	github.com/chenzhuoyu/base64x v0.0.0-20230717121745-296ad89f973d // indirect
+	github.com/chenzhuoyu/iasm v0.9.0 // indirect
+	github.com/klauspost/cpuid/v2 v2.0.9 // indirect
+	github.com/twitchyliquid64/golang-asm v0.15.1 // indirect
+	golang.org/x/arch v0.0.0-20210923205945-b76863e36670 // indirect
+)