Take a look at this, should perform even faster on amd64 because of SIMD. Sonic uses default json package as fallback but i switch on goarch to make it use the faster package instead.