@jinyus 2023-09-28
This is 3x slower for me (Ryzen 7 PRO 6850U)

old:
```
Benchmark 1: ./related_concurrent
Processing time (w/o IO) 5.751708ms
Processing time (w/o IO) 7.25925ms
Processing time (w/o IO) 7.029399ms
Processing time (w/o IO) 5.833632ms
Processing time (w/o IO) 5.297594ms
Processing time (w/o IO) 5.037711ms
Processing time (w/o IO) 5.937139ms
Processing time (w/o IO) 6.12243ms
Processing time (w/o IO) 6.404243ms
Processing time (w/o IO) 6.15784ms
Processing time (w/o IO) 5.875189ms
Processing time (w/o IO) 6.771123ms
Processing time (w/o IO) 6.986935ms
  Time (mean ± σ):      29.0 ms ±   1.0 ms    [User: 58.1 ms, System: 18.5 ms]
  Range (min … max):    27.4 ms …  30.8 ms    10 runs
```

new:

```
Benchmark 1: ./related_concurrent
Processing time (w/o IO) 13.895647ms
Processing time (w/o IO) 12.15169ms
Processing time (w/o IO) 14.669079ms
Processing time (w/o IO) 14.748908ms
Processing time (w/o IO) 15.334604ms
Processing time (w/o IO) 15.951659ms
Processing time (w/o IO) 15.904096ms
Processing time (w/o IO) 17.219992ms
Processing time (w/o IO) 14.29654ms
Processing time (w/o IO) 15.600283ms
Processing time (w/o IO) 15.251771ms
Processing time (w/o IO) 13.806319ms
Processing time (w/o IO) 16.286133ms
  Time (mean ± σ):      36.2 ms ±   2.0 ms    [User: 107.7 ms, System: 41.2 ms]
  Range (min … max):    32.7 ms …  39.5 ms    10 runs
```