@DrBlury 2023-09-28
Hmm would you mind checking just the faster json package instead? You can remove the if/else that checks for go arch amd64 and remove amd64 path. 
I tried to make the whole problem faster, not the processing itself. Processing itself should be a little slower while the runtime is reduced by 5-10ms on my machine.