@jinyus 2023-09-28
That reduced it by 5ms but the original is still faster. 

new:

```
Benchmark 1: ./related_concurrent
Processing time (w/o IO) 12.111478ms
Processing time (w/o IO) 11.807313ms
Processing time (w/o IO) 12.843009ms
Processing time (w/o IO) 12.436945ms
Processing time (w/o IO) 14.371375ms
Processing time (w/o IO) 13.884852ms
Processing time (w/o IO) 11.958103ms
Processing time (w/o IO) 12.408519ms
Processing time (w/o IO) 12.780919ms
Processing time (w/o IO) 11.720988ms
Processing time (w/o IO) 12.906985ms
Processing time (w/o IO) 11.679151ms
Processing time (w/o IO) 14.547797ms
  Time (mean ± σ):      31.6 ms ±   1.5 ms    [User: 76.8 ms, System: 46.4 ms]
  Range (min … max):    30.1 ms …  34.6 ms    10 runs

```